Add 404 error handling. Split index.js into app and server. Create first test
