+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon index.js",
+    "start": "nodemon server.js",
     "seed": "node seeds/index.js"
   },
   "repository": {
@@ -32,5 +32,9 @@
     "nodemon": "^2.0.12",
     "read-excel-file": "^5.2.20",
     "sequelize": "^6.6.5"
+  },
+  "devDependencies": {
+    "jest": "^27.4.3",
+    "supertest": "^6.1.6"
   }
 }
